+++ hosts/sigma/hardware-configuration.nix
@@ -1,4 +1,4 @@
-{ config, lib, ... }:
+{ config, ... }:
 
 {
   boot.initrd.availableKernelModules = [ "xhci_pci" "thunderbolt" "nvme" ];
@@ -85,6 +85,4 @@
       ExecStart = [ "${config.boot.zfs.package}/sbin/zfs mount -a -o remount" ];
     };
   };
-
-  nixpkgs.hostPlatform = lib.mkDefault "x86_64-linux";
 }